Check in usage of the existing data sources instead of using DataChartSharedData
