+++ samples/charts/data-chart/axis-settings/Pages/DataChartAxisSettings.razor
@@ -74,7 +74,7 @@
 
 @code {
 
-    private List<Energy> DataSource;
+    private List<EnergyProductionData.Energy> DataSource;
 
     private Legend legend;
     private Legend LegendRef
@@ -97,6 +97,6 @@
         DataChartInteractivityModule.Register(IgniteUIBlazor);
         LegendModule.Register(IgniteUIBlazor);
 
-        this.DataSource = DataChartSharedData.GetEnergyProduction();
+        this.DataSource = EnergyProductionData.Generate();
     }
 }
