+++ samples/charts/data-chart/chart-titles/Pages/DataChartTitles.razor
@@ -39,7 +39,7 @@
 
 @code {
 
-    private List<Energy> DataSource;
+    private List<EnergyProductionData.Energy> DataSource;
 
     private Legend legend;
     private Legend LegendRef
@@ -62,6 +62,6 @@
         DataChartInteractivityModule.Register(IgniteUIBlazor);
         LegendModule.Register(IgniteUIBlazor);
 
-        this.DataSource = DataChartSharedData.GetEnergyProduction();
+        this.DataSource = EnergyProductionData.Generate();
     }
 }
