+++ samples/charts/data-chart/legends/Pages/DataChartLegend.razor
@@ -36,7 +36,7 @@
 
 @code {
 
-    private List<Energy> DataSource;
+    private List<EnergyProductionData.Energy> DataSource;
 
     private Legend legend;
     private Legend LegendRef
@@ -59,6 +59,6 @@
         DataChartInteractivityModule.Register(IgniteUIBlazor);
         LegendModule.Register(IgniteUIBlazor);
 
-        this.DataSource = DataChartSharedData.GetEnergyProduction();
+        this.DataSource = EnergyProductionData.Generate();
     }
 }
