+++ samples/charts/data-chart/legends/Services/EnergyProductionData.cs
@@ -3,11 +3,20 @@ using System.Collections.Generic;
 
 namespace Infragistics.Samples
 {
-    public class DataChartSharedData
+    public static class EnergyProductionData
     {
-        public static List<Energy> GetEnergyProduction() {
-            var dp = new Energy { Country = "", Coal = 400000000};
+        public class Energy
+        {
+            public string Country { get; set; }
+            public double Coal { get; set; }
+            public double Oil { get; set; }
+            public double Gas { get; set; }
+            public double Nuclear { get; set; }
+            public double Hydro { get; set; }
+        }
 
+        public static List<Energy> Generate()
+        {
             var data = new List<Energy>() {
                 new Energy { Country = "Canada", Coal = 400000000, Oil = 100000000, Gas = 175000000, Nuclear = 225000000, Hydro = 350000000 },
                 new Energy { Country = "China", Coal = 925000000, Oil = 200000000, Gas = 350000000, Nuclear = 400000000, Hydro = 625000000 },
@@ -17,16 +26,7 @@ namespace Infragistics.Samples
                 new Energy { Country = "France", Coal = 375000000, Oil = 150000000, Gas = 350000000, Nuclear = 275000000, Hydro = 325000000 }
             };
             return data;
-        }      
-    }
+        }
 
-    public class Energy
-    {
-        public string Country { get; set; }
-        public double Coal { get; set; }
-        public double Oil { get; set; }
-        public double Gas { get; set; }
-        public double Nuclear { get; set; }
-        public double Hydro { get; set; }
     }
 }
