+++ samples/charts/data-chart/series-markers/Pages/DataChartSeriesMarkers.razor
@@ -47,11 +47,24 @@
     protected override void OnInitialized()
     {
         base.OnInitialized();
-                
+
         DataChartCoreModule.Register(IgniteUIBlazor);
-        DataChartInteractivityModule.Register(IgniteUIBlazor);        
+        DataChartInteractivityModule.Register(IgniteUIBlazor);
         DataChartCategoryModule.Register(IgniteUIBlazor);
-        
-        this.DataSource = DataChartSharedData.GetOlympicMedals();        
+
+        this.DataSource = this.GenerateData();
+    }
+
+    private List<OlympicMedals> GenerateData() {
+
+        var olympicMedals = new List<OlympicMedals>();
+        olympicMedals.Add(new OlympicMedals() { Year = "1996", USA = 148, CHN = 110, RUS = 95 });
+        olympicMedals.Add(new OlympicMedals() { Year = "2000", USA = 142, CHN = 115, RUS = 91 });
+        olympicMedals.Add(new OlympicMedals() { Year = "2004", USA = 134, CHN = 121, RUS = 86 });
+        olympicMedals.Add(new OlympicMedals() { Year = "2008", USA = 131, CHN = 129, RUS = 65 });
+        olympicMedals.Add(new OlympicMedals() { Year = "2012", USA = 135, CHN = 115, RUS = 77 });
+        olympicMedals.Add(new OlympicMedals() { Year = "2016", USA = 146, CHN = 112, RUS = 88 });
+
+        return olympicMedals;
     }
 }
